// src/interfaces/api.types.generated.d.ts
export enum CarMotorEnum {
  Electric = 'Electric',
  Hybrid = 'Hybrid'
}
// ...
export enum CarRangeEnum {
  UpTo_200 = 'UP_TO_200',
  UpTo_300 = 'UP_TO_300',
  UpTo_400 = 'UP_TO_400'
}
// ...
export enum CarTransmissionTypeEnum {
  Automatic = 'Automatic',
  Manual = 'Manual'
}
// ...
export enum CarTypeEnum {
  Car = 'Car',
  Van = 'Van'
}

export enum CarValueEnum {
  Between10And20 = 'Between10And20',
  Between20And30 = 'Between20And30',
  Between30And40 = 'Between30And40',
  Between40And60 = 'Between40And60',
  Between60And75 = 'Between60And75',
  MoreThan75 = 'MoreThan75'
}