// src/hooks/api/predictCarDetails/predictCarDetails.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Scalar for receiving email from client */
  Email: any;
};
// ...
export enum CarMotorEnum {
  Electric = 'Electric',
  Hybrid = 'Hybrid'
}
// ...
export enum CarTransmissionTypeEnum {
  Automatic = 'Automatic',
  Manual = 'Manual'
}
// ...
export type Mutation = {
// ...
  predictCarDetails?: Maybe<PredictCarDetailsResultType>;
// ...
};
// ...
export type PredictCarDetailsResultType = {
  __typename?: 'PredictCarDetailsResultType';
  color?: Maybe<Scalars['String']>;
  doorCount?: Maybe<Scalars['Float']>;
  make?: Maybe<Scalars['String']>;
  model?: Maybe<Scalars['String']>;
  motor?: Maybe<CarMotorEnum>;
  seatCount?: Maybe<Scalars['Float']>;
  transmission?: Maybe<CarTransmissionTypeEnum>;
  value?: Maybe<Scalars['Float']>;
  year?: Maybe<Scalars['Float']>;
};
// ...
export type PredictCarDetailsMutation = { __typename?: 'Mutation', predictCarDetails?: { __typename?: 'PredictCarDetailsResultType', make?: string | null, model?: string | null, year?: number | null, color?: string | null, value?: number | null, doorCount?: number | null, seatCount?: number | null, transmission?: Types.CarTransmissionTypeEnum | null, motor?: Types.CarMotorEnum | null } | null };